Static typing fix

+++ fannypack/scripts/_buddy_cli_utils.py
@@ -41,7 +41,7 @@ def get_size(path: str) -> int:
         return 0
 
 
-def format_size(size: int, short: bool=False):
+def format_size(size: float, short: bool=False):
     """Converts a size in bytes to a human-readable string.
     """
     units = ["B", "KiB", "MiB", "GiB", "TiB", "PiB"]
